+++ AntraShare/src/app/app.component.html
@@ -1,4 +1,6 @@
 <!-- <app-login></app-login> -->
 <!-- <app-setting></app-setting> -->
-<app-admin></app-admin>
+<!-- <app-admin></app-admin> -->
 <!-- <app-register></app-register> -->
+<!-- <app-news-feed></app-news-feed> -->
+<app-profile></app-profile>
